requirements.txt
aiogram==3.26.0
pillow